+++ app/src/components/ai/index.ts
@@ -2,3 +2,8 @@ export * from "./attachment";
 export * from "./elicitation";
 export * from "./prompt-input";
 export * from "./shimmer";
+// Note: `message` and `token-usage` are deliberately not re-exported here —
+// their components import from the top-level `@phoenix/components` barrel
+// (which re-exports this file), so surfacing them would close a circular
+// dependency. Import them by directory instead, e.g.
+// `@phoenix/components/ai/token-usage`.
